+++ CHANGELOG.md
@@ -11,6 +11,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Removed
 - On Interact Void event on the Interactable
     - During development, the *Interactor* parameter was added to the On Interact event. The On Interact Void event is a relic of development that has no reason to exist.
+    
+- Interactor debug line
+    - Left over from development
 
 
 ## [1.2.0] - Selfish Colors Update - 06-10-2022